requirements/base.txt
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.1.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
asyncio==3.4.3
whitenoise==6.6.0  # https://github.com/evansd/whitenoise
redis==5.0.1  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
uvicorn[standard]==0.24.0.post1  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==4.2.8  # pyup: < 5.0  # https://www.djangoproject.com/
django-environ==0.11.2  # https://github.com/joke2k/django-environ
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
django-redis==5.4.0  # https://github.com/jazzband/django-redis

# Django Admin
# ------------------------------------------------------------------------------
django-admin-honeypot==1.1.0  # https://pypi.org/project/django-admin-honeypot/
django-4-jet==1.0.9
feedparser==6.0.10
google-api-python-client==1.4.1

# Image processing
# ------------------------------------------------------------------------------
python-magic==0.4.27
imageio==2.33.1  # can read byte images and shows them to the user

# log informations
# ------------------------------------------------------------------------------
loguru==0.7.2

# Text formatter
# ------------------------------------------------------------------------------
thefuzz[speedup]==0.20.0

# Chat GPT
# ------------------------------------------------------------------------------
openai==1.7.2

# Telegram
# ------------------------------------------------------------------------------
python-telegram-bot==20.4


# Web3 & Crypto
# ------------------------------------------------------------------------------
web3==6.9.0
uniswap-python==0.7.1
cctx==1.0.0
uniswap_universal_router_decoder==0.9.1
coinbase==2.1.0
coinmarketcap==5.0.3


# Profanity check
alt-profanity-check==1.3.2
